refactor socialIcons.tsx
