+++ packages/apps/frontera/src/routes/organization/src/components/Tabs/shared/SocialMediaList/SocialIcons.tsx
@@ -1,29 +1,30 @@
 import { cn } from '@ui/utils/cn';
-import { Logo } from '@ui/media/Logo';
+import { Logo, LogoName } from '@ui/media/Logo/Logo';
 
 import { isKnownUrl } from './util';
 
-export const SocialIcon = ({
-  children,
-  className,
-  url,
-}: React.PropsWithChildren<{ url: string; className?: string }>) => {
-  const knownUrl = isKnownUrl(url);
+const logoConfig: Partial<Record<string, { fill?: string; name: LogoName }>> = {
+  twitter: { name: 'twitter' },
+  facebook: { name: 'facebook' },
+  linkedin: { name: 'linkedin', fill: '#0A66C2' },
+  github: { name: 'github' },
+  instagram: { name: 'instagram' },
+  youtube: { name: 'youtube' },
+  google: { name: 'google' },
+};
+interface SocialIconProps {
+  url: string;
+  className?: string;
+  children?: React.ReactNode;
+}
 
-  if (knownUrl === 'twitter')
-    return <Logo name='twitter' className={cn(className)} />;
-  if (knownUrl === 'facebook')
-    return <Logo name='facebook' className={cn(className)} />;
-  if (knownUrl === 'linkedin')
-    return <Logo fill='#0A66C2' name='linkedin' className={cn(className)} />;
-  if (knownUrl === 'github')
-    return <Logo name='github' className={cn(className)} />;
-  if (knownUrl === 'instagram')
-    return <Logo name='instagram' className={cn(className)} />;
-  if (knownUrl === 'youtube')
-    return <Logo name='youtube' className={cn(className)} />;
-  if (knownUrl === 'google')
-    return <Logo name='google' className={cn('size-4', className)} />;
+export const SocialIcon = ({ children, className, url }: SocialIconProps) => {
+  const knownUrl = isKnownUrl(url);
+  const logoProps = logoConfig[knownUrl as keyof typeof logoConfig];
+
+  if (logoProps) {
+    return <Logo {...logoProps} className={cn(className)} />;
+  }
 
   if (children) {
     return <>{children}</>;
